chore(structure): Move s3 integration test to integration tests
